Add dataman_s4_load_ram.py
This is not the way it should be done, but I got frustrated when dms4util.py
wasn't working so succumbed to temptation and wrote another program from
scratch to download a file to the Dataman S4 RAM.  It seems to work better at
this time and at some point, maybe I'll actually get around to resolving them
back into a single program that is better.  My apologies.